[OB-14] Refactor "header" feature
