--- src/Features/components/cities/index.ts
@@ -1,1 +1,0 @@
-export { Cities } from './cities';
