+++ src/Features/header/components/Cities/index.ts
@@ -1,0 +1,1 @@
+export { Cities } from './Cities';
