+++ src/Features/header/components/Sidebar/index.ts
@@ -1,0 +1,1 @@
+export { Sidebar } from './Sidebar';
