+++ src/Features/header/index.ts
@@ -1,3 +1,3 @@
-import { HeaderPage } from './header';
-
-export { HeaderPage };
+export * from './header';
+export { default as headerReducer } from './headerSlice';
+export * from './headerSlice';
