--- src/Features/sidebar/index.ts
@@ -1,1 +1,0 @@
-export { Sidebar } from './sidebar';
